Setup: Check if the packages are installed before installing

+++ setup.sh
@@ -1,12 +1,14 @@
 #!/usr/bin/env bash
-DEBIAN_FRONTEND=noninteractive
+export DEBIAN_FRONTEND=noninteractive
 
 function update_apt() {
+    echo "Updating apt..."
     apt-get update
     apt autoremove -y
 }
 
 function setup_prerequisite() {
+    echo "Installing pre-requisites..."
     # Adding software-properties-common for add-apt-repository.
     apt install -y software-properties-common
     # Adding git for cloning our repo
@@ -23,15 +25,18 @@ function setup_prerequisite() {
 }
 
 function setup_composer() {
-    php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-    php composer-setup.php --quiet
-    rm composer-setup.php
-    mv composer.phar /usr/local/bin/composer
+    if ! command -v composer >/dev/null 2>&1; then
+        echo "Installing Composer..."
+        php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+        php composer-setup.php --quiet
+        rm composer-setup.php
+        mv composer.phar /usr/local/bin/composer
+    fi
 }
 
 function setup_php() {
     if ! command -v php >/dev/null 2>&1; then
-        echo "Installing PHP CLI"
+        echo "Installing PHP CLI..."
         # Adding ondrej/php repository for installing php, this works for all ubuntu flavours.
         add-apt-repository -y ppa:ondrej/php
         apt-get update
@@ -41,14 +46,18 @@ function setup_php() {
 }
 
 function install_bediq_cli() {
-    echo "Downloading bedIQ CLI"
-    git clone https://github.com/bedIQ/bediq-cli.git /opt/bediq-cli
-    ln -s /opt/bediq-cli/bediq /usr/local/bin/bediq
-    cd /opt/bediq-cli
-    composer install --no-interaction --prefer-dist --optimize-autoloader
+    if ! command -v bediq >/dev/null 2>&1; then
+        echo "Downloading bedIQ CLI"
+        git clone https://github.com/bedIQ/bediq-cli.git /opt/bediq-cli
+        ln -s /opt/bediq-cli/bediq /usr/local/bin/bediq
+        cd /opt/bediq-cli
+        composer install --no-interaction --prefer-dist --optimize-autoloader
+        echo "bedIQ CLI installed"
+    fi
 }
 
 function provision_vm() {
+    echo "Starting bediq provision:vm ..."
     bediq provision:vm
 }
